Rename gan -> trainer
